+++ fhmake/__main__.py
@@ -1,4 +1,4 @@
-""" entry point for python -m fhmake """
+"""Entry point for python -m fhmake."""
 from . import cli
 
 cli()