+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2021.1 - 2021/03/28
+- Significant refactoring
+- Add score + duplicate to audit (similar to functions provided by codacy)
+
+
 ## 2021 - 2021/03/01
 - Bugfix requirements.txt generation.